2019.6.19

+++ leyou/ly-item/ly-item-pojo/src/main/java/com/leyou/dto/BrandDTO.java
@@ -2,14 +2,18 @@ package com.leyou.dto;
 
 import lombok.Data;
 
+import java.util.List;
+
 /**
  * @author 品牌DTO
  */
 
 @Data
 public class BrandDTO {
-    private Integer id;
+    private Long id;
     private String name;
     private String image;
     private Character letter;
+
+
 }
